chore: Mantenimiento del sofware y eliminacion de algunos archivos
